Fixing prometheus community repo name
The prometheus community repo name generated a cronjob with a
name that is too long by 2 characters. This shortens up the name
so the cronjob can be created.

Note, this issue has been improved in upstream monocular. It will
not be fixed in the Helm Hub as we are deprecating the Helm Hub.

Signed-off-by: Matt Farina <matt@mattfarina.com>